Traslado de las funciones HazOpeacion y ResolverPedido a hacia la biblioteca de clases Protocolo
