+++ Cliente/Form1.cs
@@ -9,32 +9,35 @@ namespace Cliente
 {
     public partial class FrmValidador : Form
     {
+
+        // Variables de clase para el cliente TCP y el flujo de datos.
         private TcpClient remoto;
         private NetworkStream flujo;
 
+        // Constructor de la clase, inicializa los componentes del formulario.
         public FrmValidador()
         {
             InitializeComponent();
         }
 
+        // Método que se ejecuta cuando el formulario se carga.
         private void FrmValidador_Load(object sender, EventArgs e)
         {
             try
             {
+                // Intenta conectar al servidor TCP en localhost en el puerto 8080.
                 remoto = new TcpClient("127.0.0.1", 8080);
                 flujo = remoto.GetStream();
             }
             catch (SocketException ex)
             {
-                MessageBox.Show("No se puedo establecer conexión " + ex.Message,
-                    "ERROR");
-            }
-            finally 
-            {
+                // Manejo de excepciones en caso de error al conectar.
+                MessageBox.Show("No se puedo establecer conexión " + ex.Message, "ERROR");
                 flujo?.Close();
                 remoto?.Close();
             }
 
+            // Desactiva los paneles y checkboxes al inicio.
             panPlaca.Enabled = false;
             chkLunes.Enabled = false;
             chkMartes.Enabled = false;
@@ -45,14 +48,15 @@ namespace Cliente
             chkSabado.Enabled = false;
         }
 
+        // Método que se ejecuta cuando se hace clic en el botón "Iniciar".
         private void btnIniciar_Click(object sender, EventArgs e)
         {
             string usuario = txtUsuario.Text;
             string contraseña = txtPassword.Text;
             if (usuario == "" || contraseña == "")
             {
-                MessageBox.Show("Se requiere el ingreso de usuario y contraseña",
-                    "ADVERTENCIA");
+                // Verifica que ambos campos estén llenos.
+                MessageBox.Show("Se requiere el ingreso de usuario y contraseña", "ADVERTENCIA");
                 return;
             }
 
@@ -61,14 +65,9 @@ namespace Cliente
                 Comando = "INGRESO",
                 Parametros = new[] { usuario, contraseña }
             };
-            
-            Respuesta respuesta = HazOperacion(pedido);
-            if (respuesta == null)
-            {
-                MessageBox.Show("Hubo un error", "ERROR");
-                return;
-            }
+            Respuesta respuesta = Protocolo.Protocolo.HazOperacion(remoto, pedido);
 
+            // Procesa la respuesta del servidor.
             if (respuesta.Estado == "OK" && respuesta.Mensaje == "ACCESO_CONCEDIDO")
             {
                 panPlaca.Enabled = true;
@@ -80,72 +79,27 @@ namespace Cliente
             {
                 panPlaca.Enabled = false;
                 panLogin.Enabled = true;
-                MessageBox.Show("No se pudo ingresar, revise credenciales",
-                    "ERROR");
+                MessageBox.Show("No se pudo ingresar, revise credenciales", "ERROR");
                 txtUsuario.Focus();
             }
         }
 
-        private Respuesta HazOperacion(Pedido pedido)
-        {
-            if(flujo == null)
-            {
-                MessageBox.Show("No hay conexión", "ERROR");
-                return null;
-            }
-            try
-            {
-                byte[] bufferTx = Encoding.UTF8.GetBytes(
-                    pedido.Comando + " " + string.Join(" ", pedido.Parametros));
-                
-                flujo.Write(bufferTx, 0, bufferTx.Length);
-
-                byte[] bufferRx = new byte[1024];
-                
-                int bytesRx = flujo.Read(bufferRx, 0, bufferRx.Length);
-                
-                string mensaje = Encoding.UTF8.GetString(bufferRx, 0, bytesRx);
-                
-                var partes = mensaje.Split(' ');
-                
-                return new Respuesta
-                {
-                    Estado = partes[0],
-                    Mensaje = string.Join(" ", partes.Skip(1).ToArray())
-                };
-            }
-            catch (SocketException ex)
-            {
-                MessageBox.Show("Error al intentar transmitir " + ex.Message,
-                    "ERROR");
-            }
-            finally 
-            {
-                flujo?.Close();
-                remoto?.Close();
-            }
-            return null;
-        }
-
+        // Método que se ejecuta cuando se hace clic en el botón "Consultar".
         private void btnConsultar_Click(object sender, EventArgs e)
         {
             string modelo = txtModelo.Text;
             string marca = txtMarca.Text;
             string placa = txtPlaca.Text;
-            
+
+            // Crea un nuevo pedido de cálculo.
             Pedido pedido = new Pedido
             {
                 Comando = "CALCULO",
                 Parametros = new[] { modelo, marca, placa }
             };
-            
-            Respuesta respuesta = HazOperacion(pedido);
-            if (respuesta == null)
-            {
-                MessageBox.Show("Hubo un error", "ERROR");
-                return;
-            }
+            Respuesta respuesta = Protocolo.Protocolo.HazOperacion(remoto, pedido);
 
+            // Procesa la respuesta del servidor.
             if (respuesta.Estado == "NOK")
             {
                 MessageBox.Show("Error en la solicitud.", "ERROR");
@@ -158,8 +112,7 @@ namespace Cliente
             else
             {
                 var partes = respuesta.Mensaje.Split(' ');
-                MessageBox.Show("Se recibió: " + respuesta.Mensaje,
-                    "INFORMACIÓN");
+                MessageBox.Show("Se recibió: " + respuesta.Mensaje, "INFORMACIÓN");
                 byte resultado = Byte.Parse(partes[1]);
                 switch (resultado)
                 {
@@ -209,36 +162,32 @@ namespace Cliente
             }
         }
 
+        // Método que se ejecuta cuando se hace clic en el botón "Número de Consultas".
         private void btnNumConsultas_Click(object sender, EventArgs e)
         {
-            String mensaje = "hola";
-            
+            //String mensaje = "hola";
+
+            // Crea un nuevo pedido para obtener el contador.
             Pedido pedido = new Pedido
             {
                 Comando = "CONTADOR",
-                Parametros = new[] { mensaje }
+                Parametros = new[] { "hola" }
             };
+            Respuesta respuesta = Protocolo.Protocolo.HazOperacion(remoto, pedido);
 
-            Respuesta respuesta = HazOperacion(pedido);
-            if (respuesta == null)
-            {
-                MessageBox.Show("Hubo un error", "ERROR");
-                return;
-            }
-
+            // Procesa la respuesta del servidor.
             if (respuesta.Estado == "NOK")
             {
                 MessageBox.Show("Error en la solicitud.", "ERROR");
-
             }
             else
             {
                 var partes = respuesta.Mensaje.Split(' ');
-                MessageBox.Show("El número de pedidos recibidos en este cliente es " + partes[0],
-                    "INFORMACIÓN");
+                MessageBox.Show("El número de pedidos recibidos en este cliente es " + partes[0], "INFORMACIÓN");
             }
         }
 
+        // Método que se ejecuta cuando el formulario se está cerrando.
         private void FrmValidador_FormClosing(object sender, FormClosingEventArgs e)
         {
             if (flujo != null)
