+++ Servidor/Servidor.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading;
 using System.Net;
 using System.Net.Sockets;
@@ -11,60 +9,65 @@ namespace Servidor
 {
     class Servidor
     {
+        // Escuchador para aceptar conexiones TCP.
         private static TcpListener escuchador;
-        private static Dictionary<string, int> listadoClientes
-            = new Dictionary<string, int>();
 
         static void Main(string[] args)
         {
             try
             {
+                // Inicializa el escuchador para aceptar conexiones en cualquier IP en el puerto 8080.
                 escuchador = new TcpListener(IPAddress.Any, 8080);
                 escuchador.Start();
                 Console.WriteLine("Servidor inició en el puerto 5000...");
 
+                // Bucle infinito para aceptar conexiones de clientes.
                 while (true)
                 {
                     TcpClient cliente = escuchador.AcceptTcpClient();
                     Console.WriteLine("Cliente conectado, puerto: {0}", cliente.Client.RemoteEndPoint.ToString());
+                    // Crea un nuevo hilo para manejar la conexión del cliente.
                     Thread hiloCliente = new Thread(ManipuladorCliente);
                     hiloCliente.Start(cliente);
                 }
             }
             catch (SocketException ex)
             {
-                Console.WriteLine("Error de socket al iniciar el servidor: " +
-                    ex.Message);
-            }
-            finally 
-            {
+                // Manejo de excepciones en caso de error al iniciar el servidor.
+                Console.WriteLine("Error de socket al iniciar el servidor: " + ex.Message);
                 escuchador?.Stop();
             }
         }
 
+        // Método para manejar la conexión con el cliente.
         private static void ManipuladorCliente(object obj)
         {
             TcpClient cliente = (TcpClient)obj;
             NetworkStream flujo = null;
             try
             {
+                // Obtiene el flujo de datos del cliente.
                 flujo = cliente.GetStream();
                 byte[] bufferTx;
                 byte[] bufferRx = new byte[1024];
                 int bytesRx;
 
+                // Bucle para leer los datos enviados por el cliente.
                 while ((bytesRx = flujo.Read(bufferRx, 0, bufferRx.Length)) > 0)
                 {
-                    string mensajeRx =
-                        Encoding.UTF8.GetString(bufferRx, 0, bytesRx);
+                    // Convierte los bytes recibidos en un string.
+                    string mensajeRx = Encoding.UTF8.GetString(bufferRx, 0, bytesRx);
+                    // Procesa el mensaje recibido para crear un objeto Pedido.
                     Pedido pedido = Pedido.Procesar(mensajeRx);
                     Console.WriteLine("Se recibio: " + pedido);
 
-                    string direccionCliente =
-                        cliente.Client.RemoteEndPoint.ToString();
-                    Respuesta respuesta = ResolverPedido(pedido, direccionCliente);
+                    // Obtiene la dirección del cliente.
+                    string direccionCliente = cliente.Client.RemoteEndPoint.ToString();
+                    // Resuelve el pedido y obtiene una respuesta.
+                    Respuesta respuesta = Protocolo.Protocolo.ResolverPedido(pedido, direccionCliente);
                     Console.WriteLine("Se envió: " + respuesta);
 
+                    // Envía la respuesta al cliente.
                     bufferTx = Encoding.UTF8.GetBytes(respuesta.ToString());
                     flujo.Write(bufferTx, 0, bufferTx.Length);
                 }
@@ -72,120 +75,12 @@ namespace Servidor
             }
             catch (SocketException ex)
             {
+                // Manejo de excepciones en caso de error al manejar la conexión del cliente.
                 Console.WriteLine("Error de socket al manejar el cliente: " + ex.Message);
-            }
-            finally
-            {
                 flujo?.Close();
                 cliente?.Close();
             }
+
         }
-
-        private static Respuesta ResolverPedido(Pedido pedido, string direccionCliente)
-        {
-            Respuesta respuesta = new Respuesta
-            { Estado = "NOK", Mensaje = "Comando no reconocido" };
-
-            switch (pedido.Comando)
-            {
-                case "INGRESO":
-                    if (pedido.Parametros.Length == 2 &&
-                        pedido.Parametros[0] == "root" &&
-                        pedido.Parametros[1] == "admin20")
-                    {
-                        respuesta = new Random().Next(2) == 0
-                            ? new Respuesta 
-                            { Estado = "OK", 
-                                Mensaje = "ACCESO_CONCEDIDO" }
-                            : new Respuesta 
-                            { Estado = "NOK", 
-                                Mensaje = "ACCESO_NEGADO" };
-                    }
-                    else
-                    {
-                        respuesta.Mensaje = "ACCESO_NEGADO";
-                    }
-                    break;
-
-                case "CALCULO":
-                    if (pedido.Parametros.Length == 3)
-                    {
-                        string modelo = pedido.Parametros[0];
-                        string marca = pedido.Parametros[1];
-                        string placa = pedido.Parametros[2];
-                        if (ValidarPlaca(placa))
-                        {
-                            byte indicadorDia = ObtenerIndicadorDia(placa);
-                            respuesta = new Respuesta
-                            { Estado = "OK", 
-                                Mensaje = $"{placa} {indicadorDia}" };
-                            ContadorCliente(direccionCliente);
-                        }
-                        else
-                        {
-                            respuesta.Mensaje = "Placa no válida";
-                        }
-                    }
-                    break;
-
-                case "CONTADOR":
-                    if (listadoClientes.ContainsKey(direccionCliente))
-                    {
-                        respuesta = new Respuesta
-                        { Estado = "OK",
-                            Mensaje = listadoClientes[direccionCliente].ToString() };
-                    }
-                    else
-                    {
-                        respuesta.Mensaje = "No hay solicitudes previas";
-                    }
-                    break;
-            }
-
-            return respuesta;
-        }
-
-        private static bool ValidarPlaca(string placa)
-        {
-            return Regex.IsMatch(placa, @"^[A-Z]{3}[0-9]{4}$");
-        }
-
-        private static byte ObtenerIndicadorDia(string placa)
-        {
-            int ultimoDigito = int.Parse(placa.Substring(6, 1));
-            switch (ultimoDigito)
-            {
-                case 1: 
-                case 2: 
-                    return 0b00100000; // Lunes
-                case 3: 
-                case 4: 
-                    return 0b00010000; // Martes
-                case 5: 
-                case 6: 
-                    return 0b00001000; // Miércoles
-                case 7: 
-                case 8: 
-                    return 0b00000100; // Jueves
-                case 9: 
-                case 0: 
-                    return 0b00000010; // Viernes
-                default: 
-                    return 0;
-            }
-        }
-
-        private static void ContadorCliente(string direccionCliente)
-        {
-            if (listadoClientes.ContainsKey(direccionCliente))
-            {
-                listadoClientes[direccionCliente]++;
-            }
-            else
-            {
-                listadoClientes[direccionCliente] = 1;
-            }
-        }
-
     }
 }
